Revert "GUI changes, optical flow field visualization, etc."
