+++ gui/__init__.py
@@ -1,18 +1,12 @@
-from gui.pages.home_page import create_home_page
-from gui.pages.process_page import create_process_page
-from gui.pages.combine_page import create_combine_page
-from gui.frames.barcode_tab import create_barcode_frame
-from gui.frames.binarization_tab import create_binarization_frame
-from gui.frames.execution_tab import create_execution_frame
-from gui.frames.flow_tab import create_flow_frame
-from gui.frames.intensity_tab import create_intensity_frame
+from gui.barcode_tab import create_barcode_frame
+from gui.binarization_tab import create_binarization_frame
+from gui.execution_tab import create_execution_frame
+from gui.flow_tab import create_flow_frame
+from gui.intensity_tab import create_intensity_frame
 
 from gui.window import setup_main_window, setup_scrollable_container, setup_log_window
 
 __all__ = [
-    "create_home_page",
-    "create_process_page",
-    "create_combine_page"
     "create_barcode_frame",
     "create_binarization_frame",
     "create_execution_frame",
