+++ gui/barcode_tab.py
