+++ gui/config.py
@@ -6,7 +6,6 @@ Generated by: python config.py
 from dataclasses import dataclass, field
 import tkinter as tk
 from core.config import *
-from typing import List
 
 @dataclass
 class InputConfigGUI:
@@ -311,14 +310,14 @@ class AggregationConfigGUI:
     generate_barcode: tk.BooleanVar = field(init=False)
     sort_parameter: tk.StringVar = field(init=False)
     normalize_barcode: tk.BooleanVar = field(init=False)
-    csv_paths_list: List[str] = field(default_factory=list)
+    csv_paths_list: tk.StringVar = field(init=False)
 
     def __post_init__(self):
         self.output_location = tk.StringVar(value=self._core_config.output_location)
         self.generate_barcode = tk.BooleanVar(value=self._core_config.generate_barcode)
         self.sort_parameter = tk.StringVar(value=self._core_config.sort_parameter)
         self.normalize_barcode = tk.BooleanVar(value=self._core_config.normalize_barcode)
-        self.csv_paths_list = list(self._core_config.csv_paths_list)
+        self.csv_paths_list = tk.StringVar(value=self._core_config.csv_paths_list)
 
     @property
     def config(self) -> AggregationConfig:
@@ -328,7 +327,7 @@ class AggregationConfigGUI:
             generate_barcode=self.generate_barcode.get(),
             sort_parameter=self.sort_parameter.get(),
             normalize_barcode=self.normalize_barcode.get(),
-            csv_paths_list=self.csv_paths_list,
+            csv_paths_list=self.csv_paths_list.get(),
         )
 
     def update_gui(self, new_config: AggregationConfig):
