+++ gui/intensity_tab.py
