+++ gui/preview_binarization.py
